methods to funcs
